Refactor models, imports, migrations * refactor default values for required fields * recreate migrations * remove relative imports
