+++ src/apps/employees/models/__init__.py
@@ -1,3 +1,3 @@
-from .position import Position
-from .specialization import Specialization
-from .employee import Employee
+from apps.employees.models.position import Position
+from apps.employees.models.specialization import Specialization
+from apps.employees.models.employee import Employee
