+++ src/apps/employees/models/position.py
@@ -1,15 +1,13 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-
 __all__ = [
     'Position'
 ]
 
 
 class Position(models.Model):
-
-    name = models.CharField(_('name'), max_length=512, default='', unique=True)
+    name = models.CharField(_('name'), max_length=512, unique=True)
 
     def __str__(self):
         return self.name
@@ -18,4 +16,3 @@ class Position(models.Model):
         ordering = ('name',)
         verbose_name = _('position')
         verbose_name_plural = _('positions')
-
