+++ src/apps/employees/models/specialization.py
@@ -1,19 +1,17 @@
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-
 __all__ = [
     'Specialization'
 ]
 
 
 class Specialization(models.Model):
-
-    name = models.CharField(_('name'), max_length=512, default='', unique=True)
+    name = models.CharField(_('name'), max_length=512, unique=True)
 
     def __str__(self):
         return self.name
 
     class Meta:
-        ordering = ('name', )
+        ordering = ('name',)
         verbose_name = _('specialization')
