+++ src/apps/employees/schema/types.py
@@ -2,8 +2,6 @@ import graphene
 
 from graphene_django.types import DjangoObjectType
 from graphene_django_extras import DjangoFilterPaginateListField
-from graphene_django_extras.paginations import LimitOffsetGraphqlPagination
-
 
 from apps.employees import models
